Hotfix for adding new cost SKU MW00330F3
Since we recently deployed changes to CI to account for the new config structure [1,2],
we want to allow time for QE to validate these changes before moving to production.

In the meantime, there are changes [3] required to resolve a BZ (related Jira [4]).
In order to avoid rushing our CI changes to PROD, this PR will add the SKU MW00330F3
from #11 directly into our `qa` branch. From there, we'll promote to the `prod`
branch and get this into production.

Once that happens, we'll still be able to merge #11 and resolve the conflict when
we open a PR from `master` to `qa`.

[1] https://github.com/RedHatInsights/entitlements-config/pull/8
[2] https://github.com/RedHatInsights/entitlements-api-go/pull/50
[3] https://github.com/RedHatInsights/entitlements-config/pull/11
[4] https://projects.engineering.redhat.com/browse/RHCLOUD-7027